migrated 'build_action' to dedicated file
