+++ crates/wasm/src/wasm_planner.rs
@@ -1,6 +1,4 @@
-use std::str::FromStr;
-
-use anyhow::{anyhow, Context};
+use anyhow::anyhow;
 use ark_ff::UniformRand;
 use decaf377::Fq;
 use rand_core::OsRng;
@@ -9,20 +7,17 @@ use wasm_bindgen::JsValue;
 
 use penumbra_chain::params::{ChainParameters, FmdParameters};
 use penumbra_dex::swap_claim::SwapClaimPlan;
-use penumbra_keys::FullViewingKey;
 use penumbra_proto::{
     core::{
         asset::v1alpha1::{DenomMetadata, Value},
         component::fee::v1alpha1::{Fee, GasPrices},
         component::ibc::v1alpha1::Ics20Withdrawal,
         keys::v1alpha1::{Address, AddressIndex},
-        transaction::v1alpha1 as pb,
         transaction::v1alpha1::MemoPlaintext,
     },
     crypto::tct::v1alpha1::StateCommitment,
     DomainType,
 };
-use penumbra_transaction::{action::Action, plan::ActionPlan, plan::TransactionPlan, WitnessData};
 
 use crate::error::WasmResult;
 use crate::planner::Planner;
@@ -65,112 +60,6 @@ impl WasmPlanner {
         Ok(planner)
     }
 
-    /// Builds a planned [`Action`] specified by
-    /// the [`ActionPlan`] in a [`TransactionPlan`].
-    /// Arguments:
-    ///     &self: `WasmPlanner`
-    ///     transaction_plan: `TransactionPlan`
-    ///     action_plan: `ActionPlan`
-    ///     full_viewing_key: `bech32m String`,
-    ///     witness_data: `WitnessData``
-    /// Returns: `Action`
-    #[wasm_bindgen]
-    pub fn build_action(
-        &self,
-        transaction_plan: JsValue,
-        action_plan: JsValue,
-        full_viewing_key: &str,
-        witness_data: JsValue,
-    ) -> WasmResult<JsValue> {
-        utils::set_panic_hook();
-
-        let transaction_plan: TransactionPlan =
-            serde_wasm_bindgen::from_value(transaction_plan.clone())?;
-
-        let witness_data_proto: pb::WitnessData = serde_wasm_bindgen::from_value(witness_data)?;
-        let witness_data: WitnessData = witness_data_proto.try_into()?;
-
-        let action_plan: ActionPlan = serde_wasm_bindgen::from_value(action_plan)?;
-
-        let full_viewing_key: FullViewingKey = FullViewingKey::from_str(full_viewing_key)?;
-
-        let memo_key = transaction_plan.memo_plan.map(|memo_plan| memo_plan.key);
-
-        let action = match action_plan {
-            ActionPlan::Spend(spend_plan) => {
-                let spend = ActionPlan::Spend(spend_plan);
-                Some(spend.build_unauth(&full_viewing_key, &witness_data, memo_key)?)
-            }
-            ActionPlan::Output(output_plan) => {
-                let output = ActionPlan::Output(output_plan);
-                Some(output.build_unauth(&full_viewing_key, &witness_data, memo_key)?)
-            }
-
-            // TODO: Other action variants besides 'Spend' and 'Output' still require testing.
-            ActionPlan::Swap(swap_plan) => {
-                let swap = ActionPlan::Swap(swap_plan);
-                Some(swap.build_unauth(&full_viewing_key, &witness_data, memo_key)?)
-            }
-            ActionPlan::SwapClaim(swap_claim_plan) => {
-                let swap_claim = ActionPlan::SwapClaim(swap_claim_plan);
-                Some(swap_claim.build_unauth(&full_viewing_key, &witness_data, memo_key)?)
-            }
-            ActionPlan::Delegate(delegation) => Some(Action::Delegate(delegation)),
-            ActionPlan::Undelegate(undelegation) => Some(Action::Undelegate(undelegation)),
-            ActionPlan::UndelegateClaim(undelegate_claim) => {
-                let undelegate_claim = undelegate_claim.undelegate_claim();
-                Some(Action::UndelegateClaim(undelegate_claim))
-            }
-            ActionPlan::ProposalSubmit(proposal_submit) => {
-                Some(Action::ProposalSubmit(proposal_submit))
-            }
-            ActionPlan::ProposalWithdraw(proposal_withdraw) => {
-                Some(Action::ProposalWithdraw(proposal_withdraw))
-            }
-            ActionPlan::ValidatorVote(validator_vote) => {
-                Some(Action::ValidatorVote(validator_vote))
-            }
-            ActionPlan::DelegatorVote(delegator_vote) => {
-                let note_commitment = delegator_vote.staked_note.commit();
-                let auth_path = witness_data
-                    .state_commitment_proofs
-                    .get(&note_commitment)
-                    .context(format!("could not get proof for {note_commitment:?}"))?;
-
-                Some(Action::DelegatorVote(delegator_vote.delegator_vote(
-                    &full_viewing_key,
-                    [0; 64].into(),
-                    auth_path.clone(),
-                )))
-            }
-            ActionPlan::ProposalDepositClaim(proposal_deposit_claim) => {
-                Some(Action::ProposalDepositClaim(proposal_deposit_claim))
-            }
-            ActionPlan::ValidatorDefinition(validator_definition) => {
-                Some(Action::ValidatorDefinition(validator_definition))
-            }
-            ActionPlan::IbcAction(ibc_action) => Some(Action::IbcRelay(ibc_action)),
-            ActionPlan::DaoSpend(dao_spend) => Some(Action::DaoSpend(dao_spend)),
-            ActionPlan::DaoOutput(dao_output) => Some(Action::DaoOutput(dao_output)),
-            ActionPlan::DaoDeposit(dao_deposit) => Some(Action::DaoDeposit(dao_deposit)),
-            ActionPlan::PositionOpen(position_open) => Some(Action::PositionOpen(position_open)),
-            ActionPlan::PositionClose(position_close) => {
-                Some(Action::PositionClose(position_close))
-            }
-            ActionPlan::PositionWithdraw(position_withdrawn) => Some(Action::PositionWithdraw(
-                position_withdrawn.position_withdraw(),
-            )),
-            ActionPlan::Withdrawal(ics20_withdrawal) => {
-                Some(Action::Ics20Withdrawal(ics20_withdrawal))
-            }
-            // TODO: Should we handle `PositionRewardClaim`?
-            _ => None,
-        };
-
-        let action_result_proto = serde_wasm_bindgen::to_value(&Some(action))?;
-        Ok(action_result_proto)
-    }
-
     /// Public getter for the 'storage' field
     pub fn get_storage(&self) -> *const IndexedDBStorage {
         &self.storage
@@ -190,6 +79,7 @@ impl WasmPlanner {
     /// Set gas prices
     /// Arguments:
     ///     gas_prices: `GasPrices`
+    #[wasm_bindgen]
     pub fn set_gas_prices(&mut self, gas_prices: JsValue) -> WasmResult<()> {
         let gas_prices_proto: GasPrices = serde_wasm_bindgen::from_value(gas_prices)?;
         self.planner.set_gas_prices(gas_prices_proto.try_into()?);
@@ -199,6 +89,7 @@ impl WasmPlanner {
     /// Add memo to plan
     /// Arguments:
     ///     memo: `MemoPlaintext`
+    #[wasm_bindgen]
     pub fn memo(&mut self, memo: JsValue) -> WasmResult<()> {
         utils::set_panic_hook();
         let memo_proto: MemoPlaintext = serde_wasm_bindgen::from_value(memo)?;
@@ -209,6 +100,7 @@ impl WasmPlanner {
     /// Add fee to plan
     /// Arguments:
     ///     fee: `Fee`
+    #[wasm_bindgen]
     pub fn fee(&mut self, fee: JsValue) -> WasmResult<()> {
         utils::set_panic_hook();
 
@@ -222,6 +114,7 @@ impl WasmPlanner {
     /// Arguments:
     ///     value: `Value`
     ///     address: `Address`
+    #[wasm_bindgen]
     pub fn output(&mut self, value: JsValue, address: JsValue) -> WasmResult<()> {
         utils::set_panic_hook();
